sched/balance: single balance prototype
This patch just collect cpu load info, and prepare balance functions.
but no balance action now (return in balance_sds directly).

This prototype patchset is very simple and include a bit hard code.
For the single alance, the key problem is finding a algorithm to
distribute load between cpus. It is a bit similar with 'bin packing'
problem. Then I try to imitate the FFD (first fit decreasing) algorithm
to distribute the loads.

The single balance will do balance for all cpus in system.
Since it can match load among all cpus, not pull only like
current load balance, we don't need the nohz_idle_balance() function.
Also we don't need much tricks for task pin tasks.

An potential problem is that if the single balance happens on a large
system with thousand CPUs, collecting cpu info may ruin the balance cpu
cache. But since whenever do load balance for wherever sched domain,
there are some cpu cache must be stained and cost time to do load balance.
So the key judgment is that if the collected info used, if we collect
some level cpu load info, but can not conduct to a load balance, like
we are already the busiest group in domain. That is more worse.

If numa distance is quit large, or system maybe too large to frequency
balance, We add a tuning knob for system level balance interval later.

Signed-off-by: Alex Shi <alex.shi@linaro.org>